requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
sqlalchemy[asyncio]>=2.0.19
aiosqlite>=0.19.0
pydantic>=2.0.0
